fix(extract-global-constants tests): Migrate tests from __tests__ to tests directory and remove redundant cases. Consolidate valid and invalid test cases for improved clarity and maintainability.
