+++ src/tests/extract-global-constants.test.ts
@@ -1,244 +1,186 @@
+import { RuleTester } from '@typescript-eslint/utils/dist/ts-eslint';
 import { extractGlobalConstants } from '../rules/extract-global-constants';
-import { ruleTesterJsx } from '../utils/ruleTester';
 
-ruleTesterJsx.run('extract-global-constants', extractGlobalConstants, {
+const ruleTester = new RuleTester({
+  parser: require.resolve('@typescript-eslint/parser'),
+  parserOptions: {
+    ecmaVersion: 2018,
+    sourceType: 'module',
+    ecmaFeatures: {
+      jsx: true,
+    },
+  },
+});
+
+ruleTester.run('extract-global-constants', extractGlobalConstants, {
   valid: [
-    // Case 1: Constants with dependencies in React components
-    `function Component({ prop }) {
-          const dependentConst = prop + 1;
-          return <div>{dependentConst}</div>;
-        }`,
-    // Case 1b: Constants with template literals that depend on variables
-    `function Component({ width }) {
-      const params = [\`w-\${width}\`];
-      return <div>{params}</div>;
-    }`,
-    // Case 1c: Template literal constant that depends on a prop
-    `function Component({ name }) {
-      const greeting = \`Hello \${name}!\`;
-      return <div>{greeting}</div>;
-    }`,
-    // Case 1d: Array constants with dependencies
-    `function Component({ prop }) {
-          const dependentConst = [prop];
-          return <div>{dependentConst}</div>;
-        }`,
-    // Case 1e: Spread operator with dependencies
-    `function Component({ prop1,prop2 }) {
-          const dependentConst = [...prop2];
-          return <div>{dependentConst}</div>;
-        }`,
-    // Case 1f: Object constants with dependencies
-    `function Component({ prop }) {
-          const dependentConst = {a: 1,prop};
-          return <div>{dependentConst}</div>;
-        }`,
-    // Case 1g: Object spread with dependencies
-    `function Component({ prop1,prop2 }) {
-          const dependentConst = {...prop2};
-          return <div>{dependentConst}</div>;
-        }`,
-    // Case 2: Local functions in React components (should be valid)
-    `function FilterDropdownSlider({ marks }) {
-          const inverseScale = (value) => {
-            const closest = marks.reduce((a, b) => {
-              const aDiff = Math.abs(a.scaledValue - value);
-              const bDiff = Math.abs(b.scaledValue - value);
-              return bDiff < aDiff ? b : a;
-            });
-            return closest;
-          };
-          return <div onClick={inverseScale}></div>;
-        }`,
-    // Case 2a: Local async functions
-    `function FilterDropdownSlider({ marks }) {
-          const inverseScale = async (value) => {
-            const closest = marks.reduce((a, b) => {
-              const aDiff = Math.abs(a.scaledValue - value);
-              const bDiff = Math.abs(b.scaledValue - value);
-              return bDiff < aDiff ? b : a;
-            });
-            return closest;
-          };
-          return <div onClick={inverseScale}></div>;
-        }`,
-    // Case 4: Arrow function components with local functions
-    `const MyComponent = ({ data }) => {
-          const processData = (item) => {
-            return item * 2;
-          };
-          return <div>{data.map(processData)}</div>;
-        }`,
-    // Case 5: React hooks with local functions
-    `function useCustomHook() {
-          const calculateValue = (input) => {
-            return input * 2;
-          };
-          return calculateValue;
-        }`,
-    // Case 6: Constants that are already in the global scope, with dependencies
-    `const GLOBAL_CONST = 'value';
-          const OTHER_GLOBAL_CONST = GLOBAL_CONST.toUpperCase();
-          function Component({prop}) {
-            return <div>{GLOBAL_CONST}{prop}</div>;
-          }`,
-    // Case 7: Constants that are not in function components or hooks
-    `const nonComponentConst = 'value';`,
-    // Case 8: Functions as constants that are not in function components or hooks
-    'const someFunc = (input: any) => `${input}`;',
-    // Case 9: Functions as constants that are not in function components or hooks
-    'const someFunc = (input: {a?: number}) => input?.a',
-    // Case 6: Dynamic imports should be considered as having dependencies
-    `function Component() {
-      const signOutFirebase = useCallback(async () => {
-        const authImport = import('../../config/firebase-client/auth');
-        const { auth } = await authImport;
-        const firebaseAuthImport = import('firebase/auth');
-        const { signOut: signOutUser } = await firebaseAuthImport;
-        return await signOutUser(auth);
-      }, []);
-      return <div>Sign Out</div>;
-    }`,
-    // For completeness
-    `export const noFilterWithoutReturn: TSESLint.RuleModule<'unexpected', never[]> =
-      createRule({
-        create(context) {
-          return {
-            'CallExpression[callee.property.name="filter"]'(
-              node: TSESTree.CallExpression,
-            ) {
-              const callback = node.arguments[0];
-              if (callback && callback.type === 'ArrowFunctionExpression') {
-                const { body } = callback;
-                if (body.type !== 'BlockStatement') {
-                  // If the body is not a block statement, it's a direct return
-                  return;
-                }
-                if (!ASTHelpers.hasReturnStatement(body)) {
-                  context.report({
-                    node,
-                    messageId: 'unexpected',
-                  });
-                }
-              }
-            },
-          };
-        },
-        name: 'no-filter-without-return',
-        meta: {
-          type: 'problem',
-          docs: {
-            description:
-              'Disallow Array.filter callbacks without an explicit return (if part of a block statement)',
-            recommended: 'error',
-          },
-          schema: [],
-          messages: {
-            unexpected:
-              'An array filter callback with a block statement must contain a return statement',
-          },
-        },
-        defaultOptions: [],
-      });
-    `,
-    `export const testingAFunction = (input: Record<string, unknown>) => {
-      const a = input?.a;
-      return a;
-    };`,
-    `export const updateCurrentGame = https.onCall(
-      async ({
-        uid,
-        gameId,
-        playing,
-      }: UpdateCurrentGameParams): UpdateCurrentGameResponse => {
-        const gameRef = db.doc(\`Game/\${gameId}\`);
-        const gameDoc = await gameRef.get();
-        const game = gameDoc.data() as Game;
-        const userStatusDatabaseRef = realtimeDb.ref(\`/status/\${uid}\`);
-        const currentRealtimeData = await userStatusDatabaseRef
-          .get()
-          .then((snap) => snap.val());
-      });`,
-    `export class StreamSecretary {
-      constructor(private readonly req: https.Request) {}
-    
-      public async verifySignature(): Promise<number | undefined> {
-        const { headers, body } = this.req;
-        const [foo,bar] = this.req
-        const { someThing } = await this.fake(this.req)
-        try {
-          Moralis.Streams.verifySignature({
-            body,
-            signature: headers['x-signature'] as string,
-          });
-          return;
-        } catch (error) {
-          logger.error(error);
-          return 401;
+    // Should allow mutable array initialization inside functions
+    {
+      code: `
+        function Component() {
+          const items = [];
+          items.push(1);
+          return items;
         }
-      }
-    }`,
-    `class Foo {
-      private get participantsChanged(): ParticipantTeam[] {
-        if (!this.tournamentBefore || !this.tournamentBefore) {
-          return [];
+      `,
+    },
+    // Should allow mutable object initialization inside functions
+    {
+      code: `
+        function Component() {
+          const config = {};
+          config.key = 'value';
+          return config;
         }
-    
-        const { participantsAggregated: participantsBefore = [] } =
-          this.tournamentBefore!;
-        const { participantsAggregated: participantsAfter = [] } =
-          this.tournamentAfter!;
-    
-        return participantsAfter.filter((teamAfter) => {
-          const teamBefore = participantsBefore.find(
-            (team) => team.id === teamAfter.id,
-          );
-          return teamBefore && !equal(teamAfter, teamBefore);
-        });
-      }
-    }`,
-    `class Foo {
-      private readonly something: string;
-      public static bar() {
-        const baz = this.something
-        return \`2*\${baz}\`;
-      }
-    }`,
-    'function useHook() { const someExtractableFunction = (input: any) => `${input}2`}',
+      `,
+    },
+    // Should allow array initialization with methods inside functions
+    {
+      code: `
+        function Component() {
+          const items = [1, 2, 3].map(x => x * 2);
+          return items;
+        }
+      `,
+    },
+    // Should allow array initialization with spread inside functions
+    {
+      code: `
+        function Component() {
+          const items = [...someArray];
+          return items;
+        }
+      `,
+    },
+    // Should allow Set/Map initialization inside functions
+    {
+      code: `
+        function Component() {
+          const set = new Set();
+          set.add(1);
+          return set;
+        }
+      `,
+    },
+    // Test case from the bug report
+    {
+      code: `
+        const menuItemsFile = useMemo(() => {
+          const items: ReactNode[] = [];
+          if (MenuItemEdit) {
+            items.push(MenuItemEdit);
+          }
+          if (MenuItemRemove) {
+            items.push(MenuItemRemove);
+          }
+          return items;
+        }, [MenuItemEdit, MenuItemRemove]);
+      `,
+    },
+    // Should allow nested array/object initialization
+    {
+      code: `
+        function Component() {
+          const nested = { arr: [1, 2, { items: [] }] };
+          nested.arr[2].items.push(3);
+          return nested;
+        }
+      `,
+    },
+    // Should allow array/object destructuring with mutation
+    {
+      code: `
+        function Component() {
+          const { items = [] } = props;
+          items.push(1);
+          return items;
+        }
+      `,
+    },
+    // Should allow class instance creation
+    {
+      code: `
+        function Component() {
+          const instance = new MyClass();
+          instance.configure();
+          return instance;
+        }
+      `,
+    },
+    // Should allow Promise chain returning mutable values
+    {
+      code: `
+        function Component() {
+          const result = Promise.resolve([])
+            .then(arr => {
+              arr.push(1);
+              return arr;
+            });
+          return result;
+        }
+      `,
+    },
   ],
   invalid: [
-    // Case 1: Constants inside function components that don't depend on anything
-    `function Component() {
-      const independentConst = 'value';
-      return <div>{independentConst}</div>;
-    }`,
-    // Case 2: Constants inside hooks that don't depend on anything
-    `function useHook() {
-      const independentConst = 'value';
-      return independentConst;
-    }`,
-    // Case 3: Constants as functions inside hooks that don't depend on anything
-    `function useHook() {
-        if (isValid) {
-          const shouldBeGlobal = 'IF_VALID_MESSAGE';
-          return 'ERROR:' + shouldBeGlobal;
+    // Should flag immutable string constants
+    {
+      code: `
+        function Component() {
+          const MESSAGE = 'Hello';
+          return MESSAGE;
         }
-        return;
-      };`,
-    `const someFunc = (a,b) => {
-      if (a > 0) {
-          const shouldBeInvalid = 'HELLO'
-      }
-    }`,
-    // Case 4: Local constants in React components (should be valid)
-    `function FilterDropdownSlider() {
-          const localConst = 'value';
-          return <div>{localConst}</div>;
-        }`,
-  ].map((testCase) => {
-    return {
-      code: testCase,
-      errors: [{ messageId: 'extractGlobalConstants' }],
-    };
-  }),
+      `,
+      errors: [
+        {
+          messageId: 'extractGlobalConstants',
+          data: { declarationName: 'MESSAGE' },
+        },
+      ],
+    },
+    // Should flag immutable number constants
+    {
+      code: `
+        function Component() {
+          const MAX_COUNT = 100;
+          return MAX_COUNT;
+        }
+      `,
+      errors: [
+        {
+          messageId: 'extractGlobalConstants',
+          data: { declarationName: 'MAX_COUNT' },
+        },
+      ],
+    },
+    // Should flag immutable boolean constants
+    {
+      code: `
+        function Component() {
+          const ENABLED = true;
+          return ENABLED;
+        }
+      `,
+      errors: [
+        {
+          messageId: 'extractGlobalConstants',
+          data: { declarationName: 'ENABLED' },
+        },
+      ],
+    },
+    // Should flag immutable RegExp constants
+    {
+      code: `
+        function Component() {
+          const REGEX = /test/;
+          return REGEX;
+        }
+      `,
+      errors: [
+        {
+          messageId: 'extractGlobalConstants',
+          data: { declarationName: 'REGEX' },
+        },
+      ],
+    },
+  ],
 });
